refactor: move ref util logics to separate file
